% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NWWNWD.R
\name{NwdNww.default}
\alias{NwdNww.default}
\title{Wyliczanie NWD NWW}
\usage{
\method{NwdNww}{default}(x, y, ...)
}
\arguments{
\item{x}{integer variable}

\item{y}{integer variable}

\item{...}{additional arguments to be passed in the future update of this package}
}
\value{
Funkcja zwaraca w wyniku NWW oraz NWD dwoch liczb
}
\description{
Szuka najmniejszej wspolnej wielokrotnosci oraz najwiekszego wspolnego dzielnika
}
\author{
Jan Bobrowski Filip Chelmikowski
}
