% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NWWNWD.R
\name{summary.NwdNww}
\alias{summary.NwdNww}
\title{Print NWD NWW}
\usage{
\method{summary}{NwdNww}(object, ...)
}
\arguments{
\item{object}{object of the NwdNww method}

\item{...}{additional arguments to be passed in the future update of this package}
}
\value{
Funkcja zwaraca w wyniku NWW oraz NWD dwoch liczb
}
\description{
Szuka najmniejszej wspolnej wielokrotnosci oraz najwiekszego wspolnego dzielnika
}
\author{
Jan Bobrowski Filip Chelmikowski
}
