% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NWWNWD.R
\name{NwdNww}
\alias{NwdNww}
\title{Szuka najmniejszej wspolnej wielokrotnosci oraz najwiekszego wspolnego dzielnika}
\usage{
NwdNww(x, y, ...)
}
\arguments{
\item{x}{integer variable}

\item{y}{integer variable}

\item{...}{additional arguments to be passed in the future update of this package}
}
\description{
Szuka najmniejszej wspolnej wielokrotnosci oraz najwiekszego wspolnego dzielnika
}
